Seperate SendRequests from inventory
