+++ tcli/display_test.py
@@ -1,3 +1,19 @@
+# Copyright 2019 Google Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+# implied. See the License for the specific language governing
+# permissions and limitations under the License.
+
+"""Tests for tcli.display."""
+
 import unittest
 
 from tcli import display
@@ -19,3 +35,6 @@ class UnitTestDisplay(unittest.TestCase):
     self.d.setColorScheme('dark')
     self.assertEqual('dark', self.d.color_scheme)
     self.assertRaises(ValueError, self.d.setColorScheme, 'bogus')
+
+if __name__ == '__main__':
+  unittest.main()
